`f32` -> `f64` (#18)

+++ src/camera/mod.rs
@@ -9,7 +9,7 @@ use crate::ray::Ray;
 use image::RgbImage;
 use nalgebra::Vector3;
 use rand::prelude::*;
-use std::f32::consts::PI;
+use std::f64::consts::PI;
 use std::io::{self, Write};
 use std::sync::Arc;
 use std::sync::mpsc::channel;
@@ -20,29 +20,29 @@ pub struct Camera {
     pub image_width: u32,
     pub image_height: u32,
 
-    pub center: Vector3<f32>,
-    pub pixel00_loc: Vector3<f32>,
-    pub pixel_delta_u: Vector3<f32>,
-    pub pixel_delta_v: Vector3<f32>,
+    pub center: Vector3<f64>,
+    pub pixel00_loc: Vector3<f64>,
+    pub pixel_delta_u: Vector3<f64>,
+    pub pixel_delta_v: Vector3<f64>,
     pub samples: u32,
-    pub samples_scale: f32,
+    pub samples_scale: f64,
     pub max_bounces: u32,
     pub threads: usize,
 
-    pub defocus_disk_u: Vector3<f32>,
-    pub defocus_disk_v: Vector3<f32>,
-    pub defocus_angle: f32,
+    pub defocus_disk_u: Vector3<f64>,
+    pub defocus_disk_v: Vector3<f64>,
+    pub defocus_angle: f64,
 
-    pub background: Vector3<f32>,
+    pub background: Vector3<f64>,
 }
 
-fn random_in_unit_disk() -> Vector3<f32> {
+fn random_in_unit_disk() -> Vector3<f64> {
     let mut rng = rand::rng();
 
     loop {
         let p = Vector3::new(
-            rng.random_range(-1.0f32..1.0f32),
-            rng.random_range(-1.0f32..1.0f32),
+            rng.random_range(-1.0f64..1.0f64),
+            rng.random_range(-1.0f64..1.0f64),
             0.0,
         );
         if p.norm_squared() < 1.0 {
@@ -51,7 +51,7 @@ fn random_in_unit_disk() -> Vector3<f32> {
     }
 }
 
-fn linear_to_gamma(linear_component: f32) -> f32 {
+fn linear_to_gamma(linear_component: f64) -> f64 {
     if linear_component > 0.0 {
         linear_component.sqrt()
     } else {
@@ -59,7 +59,7 @@ fn linear_to_gamma(linear_component: f32) -> f32 {
     }
 }
 
-fn degrees_to_radians(degress: f32) -> f32 {
+fn degrees_to_radians(degress: f64) -> f64 {
     degress * PI / 180.0
 }
 
@@ -80,8 +80,8 @@ impl Camera {
         let focus_dist = options.focus_dist;
 
         let h = (theta / 2.0).tan();
-        let viewport_height: f32 = 2.0 * h * focus_dist;
-        let viewport_width: f32 = viewport_height * (image_width as f32 / image_height as f32);
+        let viewport_height: f64 = 2.0 * h * focus_dist;
+        let viewport_width: f64 = viewport_height * (image_width as f64 / image_height as f64);
 
         let w = (look_from - look_at).normalize();
         let u = vup.cross(&w).normalize();
@@ -90,8 +90,8 @@ impl Camera {
         let viewport_u = viewport_width * u;
         let viewport_v = viewport_height * (-v);
 
-        let pixel_delta_u = viewport_u / image_width as f32;
-        let pixel_delta_v = viewport_v / image_height as f32;
+        let pixel_delta_u = viewport_u / image_width as f64;
+        let pixel_delta_v = viewport_v / image_height as f64;
 
         let viewport_upper_left =
             center - (focus_dist * w) - (viewport_u / 2.0) - (viewport_v / 2.0);
@@ -103,7 +103,7 @@ impl Camera {
         let defocus_disk_v = v * defocus_radius;
 
         let samples = options.samples;
-        let samples_scale = 1.0 / (samples as f32);
+        let samples_scale = 1.0 / (samples as f64);
 
         let max_bounces = options.max_bounces;
         let threads = options.threads;
@@ -147,8 +147,8 @@ impl Camera {
         let total = self.image_width * self.image_height;
         let is_tty = atty::is(atty::Stream::Stdout);
         let print_at = match is_tty {
-            true => (total as f32 * 0.01) as usize,
-            false => (total as f32 * 0.05) as usize,
+            true => (total as f64 * 0.01) as usize,
+            false => (total as f64 * 0.05) as usize,
         };
         for (i, (x, y, pixel)) in rx.iter().enumerate() {
             if i % print_at == 0 {
@@ -171,7 +171,7 @@ impl Camera {
     pub fn get_pixel(&self, world: &Geometry, x: u32, y: u32) -> image::Rgb<u8> {
         let color = (0..self.samples)
             .map(|_| self.ray_color(&self.get_ray(x, y), self.max_bounces, world))
-            .sum::<Vector3<f32>>()
+            .sum::<Vector3<f64>>()
             * self.samples_scale;
 
         let intensity = Interval::new(0.0, 0.999);
@@ -183,10 +183,10 @@ impl Camera {
     }
 
     pub fn get_ray(&self, x: u32, y: u32) -> Ray {
-        let offset = math::random_box(-0.5f32..0.5f32);
+        let offset = math::random_box(-0.5f64..0.5f64);
         let pixel_sample = self.pixel00_loc
-            + (self.pixel_delta_u * (offset.x + x as f32))
-            + (self.pixel_delta_v * (offset.y + y as f32));
+            + (self.pixel_delta_u * (offset.x + x as f64))
+            + (self.pixel_delta_v * (offset.y + y as f64));
 
         let ray_origin = match self.defocus_angle <= 0.0 {
             true => self.center,
@@ -195,30 +195,30 @@ impl Camera {
         let ray_direction = pixel_sample - ray_origin;
 
         let mut rng = rand::rng();
-        let t = rng.random_range(0.0f32..1.0f32);
+        let t = rng.random_range(0.0f64..1.0f64);
 
         Ray::new(ray_origin, ray_direction, t)
     }
 
-    pub fn defocus_disk_sample(&self) -> Vector3<f32> {
+    pub fn defocus_disk_sample(&self) -> Vector3<f64> {
         let p = random_in_unit_disk();
         self.center + (p.x * self.defocus_disk_u) + (p.y * self.defocus_disk_v)
     }
 
-    pub fn ray_color(&self, ray: &Ray, depth: u32, world: &Geometry) -> Vector3<f32> {
+    pub fn ray_color(&self, ray: &Ray, depth: u32, world: &Geometry) -> Vector3<f64> {
         if depth == 0 {
             return Vector3::default();
         }
 
         let mut hit_record = HitRecord::default();
-        let interval = Interval::new(0.001, f32::INFINITY);
+        let interval = Interval::new(0.001, f64::INFINITY);
 
         if !world.hit(ray, &interval, &mut hit_record) {
             return self.background;
         }
 
         let mut scattered = Ray::default();
-        let mut attenuation = Vector3::<f32>::default();
+        let mut attenuation = Vector3::<f64>::default();
         let color_from_emission =
             hit_record
                 .material
