+++ src/material/lambertian.rs
@@ -25,7 +25,7 @@ impl Surface for Lambertian {
         &self,
         r_in: &Ray,
         record: &HitRecord,
-        attenuation: &mut Vector3<f32>,
+        attenuation: &mut Vector3<f64>,
         scattered: &mut Ray,
     ) -> bool {
         let mut scatter_direction = record.normal + math::random_normal();
@@ -40,7 +40,7 @@ impl Surface for Lambertian {
         true
     }
 
-    fn emitted(&self, _: f32, _: f32, _: Vector3<f32>) -> Vector3<f32> {
-        Vector3::<f32>::default()
+    fn emitted(&self, _: f64, _: f64, _: Vector3<f64>) -> Vector3<f64> {
+        Vector3::<f64>::default()
     }
 }
